Change reentr.pl names to be legal
This continues the process started in #23592 to change names with
leading underscores to be legal C.  See that p.r. or
4bb3572f7a1c1f3944b7f58b22b6e7a9ef5faba6 for extensive discussion.

This commit simply moves the leading underscore to be trailing
Some of these may actually be legal, and don't need to be moved, as
leading underscores are only illegal in symbols with file scope.  But it
was easier to just do a global substitute, and does no harm to move them
all to be trailing.