intuit_more: Look for more variable types.
Using \w here missed the possibility of lots of other syntaxes of
variables, like $::foo or ${foo}, that scan_ident looks for.  So call
scan_ident without first filtering what it looks at.

It is more convenient to then swap the code block that handles
punctuation variables with the code block that handles other length 1
identifiers