intuit_more: 0 length identifier means nothing more
If scan_ident indicates that $ @ & are followed by nothing that looks
like an identifier, then this isn't an expression.  It has to be a
character class or an error.

Almost anything is an identifier when 'use utf8' isn't in effect;  when
it is, non ASCII has to be an Identifier Start character following these