perlapi: Add 'apidoc_flag' capability
This is meant to easily document the flags a function accepts or
returns.

Currently this only is used for generating X<> entries for the flags,
but it allows replacing (in the next commit) the previous, clumsier,
method used that does the same thing.

And it allows for future enhancements, including generating tests by
Devel::PPPort and automatic text added by autodoc.